introduce common.py
- to be shared between plotgraph.py and the plot scripts
- for the beginning, add the is_number/is_index functions
