+++ plotvectors/common.py
@@ -1,0 +1,33 @@
+# PlotGraph - Package for Sublime Text 3 to plot selected number colums as graph.
+# Copyright (C) 2017-2020 Tibor Leupold
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+
+# http://stackoverflow.com/questions/354038/ \
+# how-do-i-check-if-a-string-is-a-number-float-in-python
+def is_number(s):
+    try:
+        float(s)
+        return True
+    except ValueError:
+        return False
+
+def is_index(l, index):
+    try:
+        l[index]
+        return True
+    except IndexError:
+        return False
+
